Fix compiler warnings across headers and tests
- Fix UB: replace variadic string& constructor with non-variadic overload in CRException
- Fix format-security and sprintf in CRX_CAPTURE_CATCH/CRX_REPORT_CATCH macros
- Fix %lu format mismatch for uint64_t in condition.h CO_DEBUG calls
- Remove const qualifier on int return types for geterrno()/line()
- Change m_linenumber to unsigned int to match constructor parameter
- Add copy assignment operator to crts (deprecated-copy fix)
- Comment out unused parameters in crtimer.h and semaphore.h
- Move msg decl inside #if block in semaphore.h P() for macOS
- Add static_cast<char> in lstring.h Obfuscate::operator()
- Remove unused variable in test_bigint256.cpp
- Fix long-to-size_t sign conversion in test_crexception.cpp

+++ Source/include/condition.h
@@ -139,8 +139,8 @@ namespace crutil {
 
          ++m_waiters;
 #if (defined(_GNU_SOURCE) && !defined(ANDROID))
-         CO_DEBUG("COND(%d)[%s,%c]: Wait (m_fired=%d,m_waiters=%d,to==%lu)\n", CRX_GETTID(),
-                  m_sem.getVerbose(),(m_broadcast?'B':'N'),m_fired,m_waiters,nsec_timeout);
+         CO_DEBUG("COND(%d)[%s,%c]: Wait (m_fired=%d,m_waiters=%d,to==%llu)\n", CRX_GETTID(),
+                  m_sem.getVerbose(),(m_broadcast?'B':'N'),m_fired,m_waiters,(unsigned long long)nsec_timeout);
 #endif
 
          m_sem.cv(this);
@@ -292,8 +292,8 @@ namespace crutil {
          pthread_setspecific(Condition::CONDKEY,this);
          pthread_cleanup_push(Condition::threadCancel, this);
 
-         CO_DEBUG("COND(%d)[%s,%c]: >>> waitFor(t=%lu) <m_fired=%d,m_waiters=%d>\n", CRX_GETTID(),
-                  m_sem.getVerbose(),(m_broadcast?'B':'N'),nsec_timeout,m_fired,m_waiters);
+         CO_DEBUG("COND(%d)[%s,%c]: >>> waitFor(t=%llu) <m_fired=%d,m_waiters=%d>\n", CRX_GETTID(),
+                  m_sem.getVerbose(),(m_broadcast?'B':'N'),(unsigned long long)nsec_timeout,m_fired,m_waiters);
 #endif
 
          if (unlikely(m_broadcast == true)) {
